+++ frontend/src/pages/Unprotected/Landing.jsx
@@ -1,6 +1,6 @@
 import { useNavigate } from "react-router-dom";
-import Button from "../components/UI/Button";
-import Hero from "../components/UI/Hero";
+import Button from "../../components/UI/Button";
+import Hero from "../../components/UI/Hero";
 const Landing = () => {
   let navigate = useNavigate();
   return (
